# api/openapi.yaml
openapi: 3.0.0
info:
  title: CTF Management API
  description: 'API for managing CTF (Capture The Flag) games, teams, users, and services.'
  contact: {}
  version: 1.0.0
servers:
  - url: 'https://ctf01d.com'
    description: Production server
  - url: 'https://staging.ctf01d.com'
    description: Staging server
  - url: 'http://localhost:4102'
    description: Local server
paths:
...
  /api/v1/games:
    get:
      tags:
        - Games
      summary: List all games
      operationId: listGames
      parameters: []
      responses:
        '200':
          description: A list of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GameResponse'
                x-content-type: application/json
        '400':
          description: Bad request
    post:
      tags:
        - Games
      summary: Create a new game
      operationId: createGame
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameRequest'
        required: true
      responses:
        '200':
          description: Game created successfully
        '400':
          description: Bad request
        '500':
          description: Failed response
  '/api/v1/games/{gameId}':
    get:
      tags:
        - Games
      summary: Get a game by ID
      operationId: getGameById
      parameters:
        - name: gameId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Detailed information of a game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameResponse'
        '400':
          description: Bad request
        '500':
          description: Failed response
    put:
      tags:
        - Games
      summary: Update a game
      operationId: updateGame
      parameters:
        - name: gameId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameRequest'
        required: true
      responses:
        '200':
          description: Game updated successfully
        '400':
          description: Failed request
          content:
            application/json:
              schema: {}
        '500':
          description: Failed response
          content:
            application/json:
              schema: {}
    delete:
      tags:
        - Games
      summary: Delete a game
      operationId: deleteGame
      parameters:
        - name: gameId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Game deleted successfully
        '400':
          description: Bad request
        '500':
          description: Failed response
  /api/v1/teams:
    get:
      tags:
        - Teams
      summary: List all teams
      operationId: listTeams
      parameters: []
      responses:
        '200':
          description: A list of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamResponse'
                x-content-type: application/json
        '400':
          description: Failed request
          content:
            application/json:
              schema: {}
        '500':
          description: Failed response
          content:
            application/json:
              schema: {}
    post:
      tags:
        - Teams
      summary: Create a new team
      operationId: createTeam
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
        required: true
      responses:
        '200':
          description: Team created successfully
        '400':
          description: Failed request
          content:
            application/json:
              schema: {}
        '500':
          description: Failed response
          content:
            application/json:
              schema: {}
  '/api/v1/teams/{teamId}':
    get:
      tags:
        - Teams
      summary: Get a team by ID
      operationId: getTeamById
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Detailed information of a team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
    put:
      tags:
        - Teams
      summary: Update a team
      operationId: updateTeam
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamRequest'
        required: true
      responses:
        '200':
          description: Team updated successfully
    delete:
      tags:
        - Teams
      summary: Delete a team
      operationId: deleteTeam
      parameters:
        - name: teamId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Team deleted successfully
  /api/v1/teams/{teamId}/users/{userId}:
    post:
      tags:
        - Teams
      summary: Connect user with team
      operationId: connectUserWithTeam
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the team
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
      responses:
        '200':
          description: Connect successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    put:
      tags:
        - Teams
      summary: Approve connected user with team lead
      operationId: approveUserWithTeam
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the team
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
      responses:
        '200':
          description: Approve successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    delete:
      tags:
        - Teams
      summary: Leave user from team
      operationId: leaveUserFromTeam
      parameters:
        - in: path
          name: teamId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the team
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
          description: The ID of the user
      responses:
        '200':
          description: Left successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/v1/games/{gameId}/results:
    post:
      tags:
        - Results
      parameters:
        - in: path
          name: gameId
          required: true
          schema:
            type: string
            format: uuid
      summary: Create a new game result
      operationId: createResult
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResultRequest'
        required: true
      responses:
        '200':
          description: Result created successfully
    get:
      tags:
        - Results
      summary: Get game result
      operationId: getResult
      parameters:
        - name: gameId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Detailed information of a result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
    put:
      tags:
        - Results
      summary: Update a result
      operationId: updateResult
      parameters:
        - in: path
          name: gameId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResultRequest'
        required: true
      responses:
        '200':
          description: Result updated successfully
  /api/v1/games/{gameId}/scoreboard:
    get:
      tags:
        - Results
      summary: Get game scoreboard
      operationId: getScoreboard
      parameters:
        - name: gameId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
            format: uuid
            example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
      responses:
        '200':
          description: Detailed information of a result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultResponse'
...
  /api/v1/universities:
    get:
      summary: Retrieves a list of universities
      description: >
        This endpoint retrieves universities. It can optionally filter
        universities that match a specific term.
      tags:
        - University
      operationId: listUniversities
      parameters:
        - in: query
          name: term
          schema:
            type: string
          description: Optional search term to filter universities by name.
          required: false
      responses:
        '200':
          description: A JSON array of universities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UniversitiesResponse'
components:
  schemas:
...
    GameRequest:
      required:
        - end_time
        - start_time
      type: object
      properties:
        start_time:
          type: string
          description: The start time of the game
          format: date-time
          example: '2024-01-23T04:56:07.000Z'
        end_time:
          type: string
          description: The end time of the game
          format: date-time
          example: '2024-01-24T04:56:07.000Z'
        description:
          type: string
          description: A brief description of the game
    GameResponse:
      required:
        - end_time
        - id
        - start_time
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the game
        start_time:
          type: string
          description: The start time of the game
          format: date-time
          example: '2024-01-23T04:56:07.000Z'
        end_time:
          type: string
          description: The end time of the game
          format: date-time
          example: '2024-01-24T04:56:07.000Z'
        description:
          type: string
          description: A brief description of the game
    ResultRequest:
      required:
        - game_id
        - rank
        - score
        - team_id
      type: object
      properties:
        team_id:
          type: string
          description: Identifier of the team this result belongs to
        game_id:
          type: string
          description: Identifier of the game this result is for
        score:
          type: integer
          description: The score achieved by the team
        rank:
          type: integer
          description: The rank achieved by the team in this game
    ResultResponse:
      required:
        - id
        - game_id
        - rank
        - score
        - team_id
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the result entry
        team_id:
          type: string
          description: Identifier of the team this result belongs to
        game_id:
          type: string
          description: Identifier of the game this result is for
        score:
          type: integer
          description: The score achieved by the team
        rank:
          type: integer
          description: The rank achieved by the team in this game
...
    TeamRequest:
      required:
        - name
        - university_id
      type: object
      properties:
        name:
          type: string
          description: Name of the team
        description:
          type: string
          description: A brief description of the team
        university_id:
          type: string
          format: uuid
          description: University or institution the team is associated with
        social_links:
          type: string
          description: JSON string containing social media links of the team
        avatar_url:
          type: string
          description: URL to the team's avatar
    TeamResponse:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the team
        name:
          type: string
          description: Name of the team
        description:
          type: string
          description: A brief description of the team
        university:
          type: string
          description: University or institution the team is associated with
        social_links:
          type: string
          description: JSON string containing social media links of the team
        avatar_url:
          type: string
          description: URL to the team's avatar
    UniversityResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
          description: The unique identifier of the university
          example: 81e17cb6-5c46-4934-b6d5-5c8356712cb6
        name:
          type: string
          description: The name of the university
          example: >-
            Анапский филиал Кубанского государственного аграрного
            университета
    UniversitiesResponse:
      type: array
      items:
        $ref: '#/components/schemas/UniversityResponse'
...
security: []